+++ src/messages/execute.rs
@@ -81,21 +81,21 @@ impl ExecuteMessage<'_, '_> {
     }
 
     /// Writes bind parameter data to the buffer.
-    fn write_bind_params(&self, buf: &mut WriteBuffer, input_only: bool) {
+    fn write_bind_params(
+        &self,
+        buf: &mut WriteBuffer,
+        bind_indexes: Vec<usize>,
+    ) {
         let binds = self.statement.binds();
         for row_index in 0..self.params.num_rows() {
             buf.write_u8(constants::TTC_MSG_TYPE_ROW_DATA);
-            for (column_index, bind_info) in binds.iter().enumerate() {
-                if (!input_only && !bind_info.is_return_bind)
-                    || bind_info.is_input_bind()
-                {
-                    self.params.write_to_buf(
-                        row_index,
-                        column_index,
-                        bind_info,
-                        buf,
-                    );
-                }
+            for column_index in bind_indexes.iter() {
+                self.params.write_to_buf(
+                    row_index,
+                    *column_index,
+                    &binds[*column_index],
+                    buf,
+                );
             }
         }
     }
@@ -261,8 +261,11 @@ impl ExecuteMessage<'_, '_> {
             self.write_define_metadata(client, buf);
         } else if has_binds {
             self.write_bind_metadata(client, buf);
-            if self.statement.has_input_binds() {
-                self.write_bind_params(buf, false);
+            let bind_indexes = self
+                .statement
+                .bind_indexes_for_execute(client.max_string_size(), false);
+            if !bind_indexes.is_empty() {
+                self.write_bind_params(buf, bind_indexes);
             }
         }
     }
@@ -291,8 +294,11 @@ impl ExecuteMessage<'_, '_> {
         buf.write_ub4(num_iters);
         buf.write_ub4(options_1);
         buf.write_ub4(options_2);
-        if self.statement.has_input_binds() {
-            self.write_bind_params(buf, true);
+        let bind_indexes = self
+            .statement
+            .bind_indexes_for_execute(client.max_string_size(), true);
+        if !bind_indexes.is_empty() {
+            self.write_bind_params(buf, bind_indexes);
         }
     }
 }