+++ doc/release_notes.md
@@ -17,6 +17,8 @@
     columns in [Row::get()](crate::Row::get()) and
     [Row::take()](crate::Row::take())
     ([issue 12](https://github.com/oracle/rust-oracledb/issues/12)).
+1.  Added support for binding long values in any order
+    ([issue 10](https://github.com/oracle/rust-oracledb/issues/10)).
 1.  Added support for the HA readiness requirements of Oracle Database 23.26.3.
 1.  Removed the ability to clone [Cursor](crate::Cursor) and [Lob](crate::Lob).
 1.  Fixed bug which caused a named binding containing a single quote to panic.